# RecipeBook/docs/swagger.yaml
basePath: /
definitions:
  helpers.APIError:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
...
  helpers.DeleteResponse:
    properties:
      message:
        type: string
    type: object
...
  helpers.RecipeCommentInput:
    properties:
      message:
        type: string
    type: object
  helpers.RecipeInput:
    properties:
      category:
        type: string
      description:
        type: string
      ingredients:
        type: string
      picture_url:
        type: string
      steos:
        type: string
      tags:
        type: string
      title:
        type: string
    type: object
...
  models.Comments:
    properties:
      created_at:
        type: string
      id:
        type: integer
      message:
        type: string
      recipe:
        $ref: '#/definitions/models.Recipes'
      recipe_id:
        type: integer
      updated_at:
        type: string
      user:
        $ref: '#/definitions/models.Users'
      user_id:
        type: integer
    type: object
...
  models.Likes:
    properties:
      created_at:
        type: string
      id:
        type: integer
      recipe:
        $ref: '#/definitions/models.Recipes'
      recipe_id:
        type: integer
      updated_at:
        type: string
      user:
        $ref: '#/definitions/models.Users'
      user_id:
        type: integer
    type: object
  models.Recipes:
    properties:
      category:
        type: string
      comments:
        items:
          $ref: '#/definitions/models.Comments'
        type: array
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      ingredients:
        type: string
      likes:
        items:
          $ref: '#/definitions/models.Likes'
        type: array
      picture_url:
        type: string
      steps:
        type: string
      tags:
        type: string
      title:
        type: string
      updated_at:
        type: string
      user:
        $ref: '#/definitions/models.Users'
      user_id:
        type: integer
    type: object
  models.Users:
    properties:
      bio:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      password:
        type: string
      recipes:
        items:
          $ref: '#/definitions/models.Recipes'
        type: array
      updated_at:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: soberkoder@swagger.io
    name: API Support
  description: This is a sample service for managing cars
  termsOfService: http://swagger.io/terms/
  title: RecipeBook API
  version: "1.0"
paths:
...
  /recipes:
    get:
      consumes:
      - application/json
      description: Get details of all recipes
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Recipes'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helpers.APIError'
      security:
      - BearerAuth: []
      summary: Get details
      tags:
      - Recipes
    post:
      consumes:
      - application/json
      description: Post details of recipe corresponding to the input Id
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: create recipe
        in: body
        name: helpers.RecipeInput
        required: true
        schema:
          $ref: '#/definitions/helpers.RecipeInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Recipes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helpers.APIError'
        "500":
          description: Server Error
          schema:
            $ref: '#/definitions/helpers.APIError'
      security:
      - BearerAuth: []
      summary: Post details for a given Id
      tags:
      - Recipes
  /recipes/{recipeId}:
    delete:
      consumes:
      - application/json
      description: Delete the order corresponding to the input id
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID of the recipe to be deleted
        in: path
        name: recipeId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helpers.DeleteResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helpers.APIError'
        "500":
          description: Server Error
          schema:
            $ref: '#/definitions/helpers.APIError'
      security:
      - BearerAuth: []
      summary: Delete recipe identified by the given id
      tags:
      - Recipes
    get:
      consumes:
      - application/json
      description: Get details recipe corresponding to the input Id
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID of the recipe
        in: path
        name: recipeId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Recipes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helpers.APIError'
        "500":
          description: Server Error
          schema:
            $ref: '#/definitions/helpers.APIError'
      security:
      - BearerAuth: []
      summary: Get details for a given Id
      tags:
      - Recipes
    put:
      consumes:
      - application/json
      description: Update the recipe corresponding to the input Id
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID of the recipe to be updated
        in: path
        name: recipeId
        required: true
        type: integer
      - description: update recipe
        in: body
        name: helpers.RecipeInput
        required: true
        schema:
          $ref: '#/definitions/helpers.RecipeInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Recipes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helpers.APIError'
        "500":
          description: Server Error
          schema:
            $ref: '#/definitions/helpers.APIError'
      security:
      - BearerAuth: []
      summary: Update recipe identified by the given Id
      tags:
      - Recipes
  /recipes/{recipeId}/comments:
    post:
      consumes:
      - application/json
      description: Post details of recipe corresponding to the input Id
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID of the recipe
        in: path
        name: recipeId
        required: true
        type: integer
      - description: create recipe comment
        in: body
        name: helpers.RecipeCommentInput
        required: true
        schema:
          $ref: '#/definitions/helpers.RecipeCommentInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Comments'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helpers.APIError'
        "500":
          description: Server Error
          schema:
            $ref: '#/definitions/helpers.APIError'
      security:
      - BearerAuth: []
      summary: Post details for a given Id
      tags:
      - Recipes
  /recipes/{recipeId}/follows:
    post:
      consumes:
      - application/json
      description: Post details of recipe corresponding to the input Id
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID of the recipe
        in: path
        name: recipeId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Likes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helpers.APIError'
        "500":
          description: Server Error
          schema:
            $ref: '#/definitions/helpers.APIError'
      security:
      - BearerAuth: []
      summary: Post details for a given Id
      tags:
      - Recipes
  /recipes/{recipeId}/likes:
    post:
      consumes:
      - application/json
      description: Post details of recipe corresponding to the input Id
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID of the recipe
        in: path
        name: recipeId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Likes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helpers.APIError'
        "500":
          description: Server Error
          schema:
            $ref: '#/definitions/helpers.APIError'
      security:
      - BearerAuth: []
      summary: Post details for a given Id
      tags:
      - Recipes
...
swagger: "2.0"